Some modem cleanups, and movement towards efficient fixed point modem implementations for platforms with slow floating point

+++ libs/spandsp/src/spandsp/private/v29rx.h
@@ -87,7 +87,7 @@ struct v29_rx_state_s
 
     /*! \brief A measure of how much mismatch there is between the real constellation,
                and the decoded symbol positions. */
-    float training_error;
+    int32_t training_error;
 
     /*! \brief The proportional part of the carrier tracking filter. */
     int32_t carrier_track_p;
